Feat: admin can delete any of the psychologists from adress book

+++ Client/src/app/psychologist/psychologist/psychologist.component.ts
@@ -1,6 +1,8 @@
 import { Component, OnInit } from '@angular/core';
 import { PsychologistService } from '../psychologist.service';
 import { MatSnackBar } from '@angular/material';
+import { AuthService } from '../../auth/auth.service';
+import { Router } from '@angular/router';
 
 @Component({
   selector: 'app-psychologist',
@@ -9,10 +11,14 @@ import { MatSnackBar } from '@angular/material';
 })
 export class PsychologistComponent implements OnInit {
 
+  user: any;
   psychologists: any[];
+  admin: boolean;
 
   constructor(private psychologistService: PsychologistService,
-               public snackBar: MatSnackBar) { }
+              public snackBar: MatSnackBar,
+              private authService: AuthService,
+              private router: Router) { }
 
   getPsychologists(): void {
     let self = this;
@@ -22,23 +28,37 @@ export class PsychologistComponent implements OnInit {
   }
   deletePsychologist(index: any): void {
     const self = this;
-    this.psychologistService.deletePsychologist(self.psychologists[index]._id).subscribe(function(res) {
-      if (res.err != null) {
-        /* if an error returned notify the user to try again */
-        self.snackBar.open('Something went wrong, please try again.', '', {
-          duration: 2500
-        });
-      } else {
-        /* everything went great!! notify the user it was a success then reload. */
-        self.snackBar.open(res.msg, '', {
-          duration: 2300
-        });
-        self.getPsychologists();
-      }
-    });
+    if (this.admin) {
+      this.psychologistService.deletePsychologist(self.psychologists[index]._id).subscribe(function(res) {
+        if (res.err != null) {
+          /* if an error returned notify the user to try again */
+          self.snackBar.open('Something went wrong, please try again.', '', {
+            duration: 2500
+          });
+        } else {
+          /* everything went great!! notify the user it was a success then reload. */
+          self.snackBar.open(res.msg, '', {
+            duration: 2300
+          });
+          self.getPsychologists();
+        }
+      });
+    } else {
+      alert('your are not an admin to do that >:(');
+    }
   }
   ngOnInit() {
-    this.getPsychologists();
-  }
+    const self = this;
+    const userDataColumns = ['isAdmin'];
+    this.authService.getUserData(userDataColumns).subscribe(function (res) {
+      self.user = res.data;
+      if (!self.user) {
+        self.router.navigate(['/']);
+      } else {
+        self.admin = self.user.isAdmin;
+      }
+      self.getPsychologists();
+    });
 
+  }
 }
